[CAMEL-13371] Isolate more xml stuff in camel-jaxp
