+++ components/camel-jclouds/pom.xml
@@ -49,6 +49,11 @@
             <artifactId>camel-support</artifactId>
         </dependency>
 
+        <dependency>
+            <groupId>org.apache.camel</groupId>
+            <artifactId>camel-jaxp</artifactId>
+        </dependency>
+
         <!-- jclouds -->
         <dependency>
             <groupId>org.apache.jclouds</groupId>
