+++ core/camel-core/src/test/java/org/apache/camel/processor/ValidatingProcessorFromSourceTest.java
@@ -17,7 +17,7 @@
 package org.apache.camel.processor;
 import java.io.File;
 
-import org.apache.camel.StringSource;
+import org.apache.camel.util.xml.StringSource;
 import org.apache.camel.support.processor.validation.ValidatingProcessor;
 import org.junit.Before;
 import org.junit.Test;
