+++ core/camel-jaxp/src/main/java/org/apache/camel/converter/jaxp/XmlConverter.java
@@ -66,10 +66,10 @@ import org.xml.sax.SAXException;
 import org.xml.sax.SAXParseException;
 import org.xml.sax.XMLReader;
 
-import org.apache.camel.BytesSource;
+import org.apache.camel.util.xml.BytesSource;
 import org.apache.camel.Converter;
 import org.apache.camel.Exchange;
-import org.apache.camel.StringSource;
+import org.apache.camel.util.xml.StringSource;
 import org.apache.camel.support.CamelContextHelper;
 import org.apache.camel.support.builder.xml.StAX2SAXSource;
 import org.apache.camel.util.IOHelper;
