+++ core/camel-jaxp/src/main/java/org/apache/camel/util/xml/StringSource.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel;
+package org.apache.camel.util.xml;
 
 import java.io.ByteArrayInputStream;
 import java.io.Externalizable;
